+++ scripts/forks/chia_install.sh
@@ -27,11 +27,11 @@ else
     echo "Installing Chia CUDA binaries on ${arch_name}..."
     cd /tmp
     if [[ "${arch_name}" = "x86_64" ]]; then
-        curl -sLJO https://download.chia.net/bladebit/alpha4.3/chia-blockchain-cuda/chia-installers-linux-deb-intel.zip.zip
+        curl -sLJO https://github.com/Chia-Network/chia-blockchain/releases/download/2.0.0-rc1/chia-blockchain-cli_2.0.0rc1-1_amd64.deb
         unzip *.zip
         apt-get install ./chia-blockchain-cli*.deb
     elif [[ "${arch_name}" = "arm64" ]]; then
-        curl -sLJO https://download.chia.net/bladebit/alpha4.3/chia-blockchain-cuda/chia-installers-linux-deb-arm64.zip.zip
+        curl -sLJO https://github.com/Chia-Network/chia-blockchain/releases/download/2.0.0-rc1/chia-blockchain-cli_2.0.0rc1-1_arm64.deb
         unzip *.zip
         apt-get install ./chia-blockchain-cli*.deb
     else